Reworking shadows

+++ sources/code/EngineCore/Rendering/DeferredRenderer.hpp
@@ -92,7 +92,7 @@ namespace Grindstone {
 
 		uint32_t width = 800;
 		uint32_t height = 600;
-		uint32_t bloomMipLevelCount = 0;
+		size_t bloomMipLevelCount = 0;
 		Grindstone::GraphicsAPI::Texture* brdfLut = nullptr;
 
 		std::vector<DeferredRendererImageSet> deferredRendererImageSets;
